+++ packages/core/src/index.ts
@@ -14,6 +14,9 @@ export type {
   SimpleSmartAccountParams,
 } from "./account/simple.js";
 export type * from "./account/types.js";
+export { type SmartAccountSigner } from "./signer/types";
+export { PrivateKeySigner } from "./signer/private-key";
+export { HdAccountSigner } from "./signer/hd-account";
 
 export {
   createPublicErc4337Client,
@@ -35,4 +38,5 @@ export {
   getChain,
   getUserOperationHash,
   resolveProperties,
+  defineReadOnly,
 } from "./utils.js";