+++ packages/core/src/provider/base.ts
@@ -17,11 +17,14 @@ import {
   isValidRequest,
   type BatchUserOperationCallData,
   type UserOperationCallData,
+  type UserOperationReceipt,
+  type UserOperationResponse,
   type UserOperationStruct,
 } from "../types.js";
 import {
   asyncPipe,
   deepHexlify,
+  defineReadOnly,
   getUserOperationHash,
   resolveProperties,
 } from "../utils.js";
@@ -38,20 +41,6 @@ import type {
 export const noOpMiddleware: AccountMiddlewareFn = async (
   struct: UserOperationStruct
 ) => struct;
-
-// borrowed from ethers.js
-function defineReadOnly<T, K extends keyof T>(
-  object: T,
-  key: K,
-  value: T[K]
-): void {
-  Object.defineProperty(object, key, {
-    enumerable: true,
-    value: value,
-    writable: false,
-  });
-}
-
 export interface SmartAccountProviderOpts {
   /**
    * The maximum number of times tot try fetching a transaction receipt before giving up
@@ -187,7 +176,7 @@ export class SmartAccountProvider<
     return await this.waitForUserOperationTransaction(hash as Hash);
   };
 
-  private async waitForUserOperationTransaction(hash: Hash): Promise<Hash> {
+  waitForUserOperationTransaction = async (hash: Hash): Promise<Hash> => {
     for (let i = 0; i < this.txMaxRetries; i++) {
       await new Promise((resolve) =>
         setTimeout(resolve, this.txRetryIntervalMs)
@@ -204,7 +193,15 @@ export class SmartAccountProvider<
     }
 
     throw new Error("Failed to find transaction for User Operation");
-  }
+  };
+
+  getUserOperationByHash = (hash: Hash): Promise<UserOperationResponse> => {
+    return this.rpcClient.getUserOperationByHash(hash);
+  };
+
+  getUserOperationReceipt = (hash: Hash): Promise<UserOperationReceipt> => {
+    return this.rpcClient.getUserOperationReceipt(hash);
+  };
 
   sendUserOperation = async (
     data: UserOperationCallData | BatchUserOperationCallData